api/requirements.txt
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.5
black==23.11.0
fastapi==0.92.0
flake8==6.1.0
jwtdown-fastapi==0.5.0
pydantic==1.10.13
pyflakes==3.1.0
pymongo==4.6.0
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
uvicorn==0.17.6
Whoosh==2.7.4